rebase after #56872

+++ vllm/model_executor/models/qwen3_vl.py
@@ -2318,7 +2318,7 @@ class Qwen3VLForConditionalGeneration(
             if self.use_data_parallel:
                 return run_dp_sharded_mrope_vision_model(
                     self.visual,
-                    pixel_values.type(self.visual.dtype),
+                    pixel_values,
                     grid_thw.tolist(),
                     rope_type="rope_3d",
                 )
@@ -2344,7 +2344,7 @@ class Qwen3VLForConditionalGeneration(
                 grid_thw_list = grid_thw.tolist()
                 return run_dp_sharded_mrope_vision_model(
                     self.visual,
-                    pixel_values_videos.type(self.visual.dtype),
+                    pixel_values_videos,
                     grid_thw_list,
                     rope_type="rope_3d",
                 )
